update task provider to have useEffect
