+++ rn/AddTaskView.js
@@ -3,14 +3,16 @@ import {Button} from 'react-native';
 import {Task} from './schemas';
 import {Overlay, Input} from 'react-native-elements';
 import {styles} from './App';
+import {useTasks} from './TasksProvider';
 
 // The AddTaskView is a button for adding tasks. When the button is pressed, an
 // overlay shows up to request user input for the new task name. When the
 // "Create" button on the overlay is pressed, the overlay closes and the new
 // task is created in the realm.
-export function AddTaskView({realm, partition}) {
+export function AddTaskView() {
   const [overlayVisible, setOverlayVisible] = useState(false);
   const [newTaskName, setNewTaskName] = useState('');
+  const {createTask} = useTasks();
   return (
     <>
       <Overlay isVisible={overlayVisible} overlayStyle={{width: '90%'}}>
@@ -25,12 +27,7 @@ export function AddTaskView({realm, partition}) {
             title="Create"
             onPress={() => {
               setOverlayVisible(false);
-              realm.write(() => {
-                realm.create(
-                  'Task',
-                  new Task({name: newTaskName || 'New Task', partition}),
-                );
-              });
+              createTask(newTaskName);
             }}
           />
         </>
