+++ rn/TaskItem.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {Text} from 'react-native';
 import {ListItem} from 'react-native-elements';
 import {Task} from './schemas';
-import {useTasks} from './TaskProvider';
+import {useTasks} from './TasksProvider';
 
 // The TaskItem represents a Task in a list.
 export function TaskItem({
